Move command line options to its own class file
