+++ Steam Desktop Authenticator/Steam Desktop Authenticator.csproj
@@ -166,6 +166,7 @@
     </Compile>
     <Compile Include="Manifest.cs" />
     <Compile Include="ManifestParseException.cs" />
+    <Compile Include="CommandLineOptions.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ResourceResponseFilter.cs" />
